Improve testing by adding an end() method to stores
Debugging current tests was a mess since previously created stores kept
running and reporting their states. With this commit, and `end()` method
has been added to the Store class that allows to stop the interval that
reports events.

This method can also be used by consumers whenever the store needs to be
disposed.